[Move] 51 Setting 관련 폴더링 변경
