+++ Kerdy/Kerdy/SettingWritten/ViewModel/SettingCommenetViewModel.swift
@@ -55,7 +55,7 @@ final class SettingCommenetViewModel {
 extension SettingCommenetViewModel {
     
     func getUserComments(id: Int) {
-//        commentManager.getUserCommnets(id: id)
+        //        commentManager.getUserCommnets(id: id)
         commentManager.getUserCommnets(id: 24)
             .subscribe(onSuccess: { response in
                 self.commentsList.accept(response)
